Add and notate changes to group 2.0.0 functions
Add test to m_normalize to show effect of normalization is invertible
by multiplying the resulting matrix by the largest element in the
previous matrix.

Add a note to the m_inverseMatrix function explaining nature of bug
which causes the effect of the function to be thrown away. This could
be fixed two ways but I'm not sure which one is better.

Add test to m_elem_truncate to show the effect of truncating. The current
test proves that the function does not crash immediately but not much
else, since integer values were used to construct the matrix in the first
place. The test added results in a matrix where each row is filled with
values equal to the row number.

It is worth noting that the inverse function **still does not work after
this commit**