Apply broken changes to m_inverseMatrix
This funciton still does not work properly, but we are on our way.

It is now giving the exact multiplicative opposite of the matrix we want
to get for the Identity case, but I am not sure about the other matrix.

It may or may not be functioning properly in terms of pointer arithmetic
and it is definitely not giving good results. I will need to keep testing
what we have and make sure the pointers are correct as well as the
arithmetic results.